moved renderFooter into common DOM functions
